+++ aurora.tf
@@ -1,20 +1,19 @@
 # --- locals ---
 
+# The configuration for Aurora clusters consists of common global settings, specified in the locals
+# section below. The aurora submodule is passed the regional clusters configuration, which contains
+# optional parameters per cluster, allowing overriding the common global settings.
+
 locals {
   # --- aurora features
   # this extracts the aurora serverless feature configuration from the features variable. 
   aurora = try(var.features.aurora_serverless, {
-    enabled          = false
-    excluded_regions = []
+    enabled = false
   })
 
-  # --- aurora regions
-  # if the aurora feature is not enabled, set aurora_regions to an empty list, otherwise
-  # it will adopt the global workbench regions.
-  # this allows specific regions to be excluded from aurora clusters
-  # in the event that aurora serverless v2 is not available in all workbench regions yet. 
-  aurora_base    = local.aurora.enabled ? var.workbench_regions : []
-  aurora_regions = setsubtract(local.aurora_base, local.aurora.excluded_regions)
+  # --- aurora clusters
+  # this extracts the aurora clusters configuration from the features variable.
+  aurora_clusters = try(local.aurora.clusters, {})
 
   # --- aurora master username
   # this sets the master username for the aurora clusters.
@@ -32,45 +31,47 @@ locals {
   # this is always true for aurora serverless v2
   aurora_iam_authentication_enabled = true
 
-  # --- restore to point in time configuration
-  # this sets the point-in-time recovery configuration for the aurora clusters.
-  aurora_restore_to_point_in_time = try(local.aurora.restore_to_point_in_time, null)
-
-  # --- migrate aurora from an older cluster
-  aurora_migrate_from_cluster = try(local.aurora.migrate_from_cluster, null)
+  # --- deletion protection
+  # global flag for all regions - can be overwritten in the features variable.
+  # for unit/regression testing, this value is set to false to allow cleanup.
+  aurora_deletion_protection = (length(var.regression_testing_assume_role_arns) > 0 || var.enable_resource_protection == false) ? false : true
 }
 
 # --- modules ---
 
 module "aurora_cluster" {
-  source = "./modules/aurora"
-
-  for_each                   = toset(local.aurora_regions)
-  region                     = each.value
+  for_each = {
+    for region, clusters in local.aurora_clusters : region => clusters
+    if contains(var.workbench_regions, region) && local.aurora.enabled
+  }
+  source                     = "./modules/aurora"
+  region                     = each.key
+  clusters                   = each.value
+  workbench_regions          = var.workbench_regions
   prefix                     = local.workbench_prefix
   postgresql_version         = local.aurora_postgresql_version
   master_username            = local.aurora_master_username
   aws_managed_password       = local.aurora_aws_managed_password
   iam_authentication_enabled = local.aurora_iam_authentication_enabled
-  restore_to_point_in_time   = local.aurora_restore_to_point_in_time
-  migrate_from_cluster       = local.aurora_migrate_from_cluster
+  deletion_protection        = local.aurora_deletion_protection
 
   # vpc
-  vpc_id                   = module.vpc[each.value].vpc_id
-  vpc_private_subnet_cidrs = module.vpc[each.value].private_subnet_cidrs
+  vpc_id                   = module.vpc[each.key].vpc_id
+  vpc_private_subnet_cidrs = module.vpc[each.key].private_subnet_cidrs
 
   # tags
-  resource_tags = local.resource_tags
-  tags          = local.tags
+  tags = merge(
+    local.tags,
+    try(each.value.additional_tags, {})
+  )
 
-  # delete protection
-  deletion_protection = (length(var.regression_testing_assume_role_arns) > 0 || var.enable_resource_protection == false) ? false : true
+  resource_tags = local.resource_tags
 }
 
 module "aurora_iam" {
-  source = "./modules/aurora-iam"
+  for_each = local.aurora.enabled ? toset(["this"]) : toset([])
 
-  for_each               = local.aurora.enabled ? toset(["this"]) : toset([])
+  source                 = "./modules/aurora-iam"
   account_id             = local.account_id
   prefix                 = local.workbench_prefix
   aurora_master_username = local.aurora_master_username